Fixed Bug in -c Option; Bugfix Release
Fixed a bug where the -c option did not check for the correct set of package management tools on FreeBSD >= 10.x due to the OS version check not occuring early enough (reported by Andrea Venturoli <ml@netfence.it>). Changed version number and release date in the script along with adding CHANGELOG notes for a v0.7.2 release.